+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.1.4
 
 - Adding alias for `maxcdn.del`.
+- Fixing a bug with calling delete on a single file as string.
 
 ### v0.1.3